Extract Windows1252 to its own file
